+++ src/main/java/tn/esprit/pibakcend/service/ICommentaire.java
@@ -1,6 +1,7 @@
 package tn.esprit.pibakcend.service;
 
 import tn.esprit.pibakcend.entities.Commentaire;
+import tn.esprit.pibakcend.entities.Like;
 
 import java.util.List;
 
@@ -8,9 +9,10 @@ public interface ICommentaire {
     Commentaire addComm(Commentaire comm);
     Commentaire updateComm(Commentaire comm);
 
-    Commentaire retrieveCommentaireById(long id );
+    Commentaire retrieveCommentaireById(Integer id );
 
     List<Commentaire> retrieveAllCommentaire();
 
-    void deleteCommentaire(long id);
+    void deleteCommentaire(Integer id);
+
 }
